Remove new 'fluent' package and replace the existing Validation. This is because vavr 1.0 doesn't need to be backward compatible, so no need to duplicate and deprecate classes.
