+++ vavr/src/main/java/io/vavr/$.java
@@ -8,6 +8,7 @@ package io.vavr;
 
 import io.vavr.control.Either;
 import io.vavr.control.Try;
+import io.vavr.control.Validation;
 import io.vavr.match.annotation.Patterns;
 import io.vavr.match.annotation.Unapply;
 import io.vavr.collection.List;
@@ -75,11 +76,8 @@ class $ {
     @Unapply
     static <T> Tuple1<Throwable> Failure(Try.Failure<T> failure) { return Tuple.of(failure.getCause()); }
 
-    // Deprecated Validation
     @Unapply
-    @Deprecated
-    static <E, T> Tuple1<T> Valid(io.vavr.control.Validation.Valid<E, T> valid) { return Tuple.of(valid.get()); }
+    static <E, T> Tuple1<T> Valid(Validation.Valid<E, T> valid) { return Tuple.of(valid.get()); }
     @Unapply
-    @Deprecated
-    static <E, T> Tuple1<E> Invalid(io.vavr.control.Validation.Invalid<E, T> invalid) { return Tuple.of(invalid.getError()); }
+    static <E, T> Tuple1<E> Invalid(Validation.Invalid<E, T> invalid) { return Tuple.of((E) invalid.getErrors()); }
 }
