--- vavr/src/main/java/io/vavr/control/fluent/package-info.java
@@ -1,6 +1,0 @@
-/**
- * A package for classes that allow fluent chaining of complex functions.
- * For example {@linkplain io.vavr.control.fluent.Validation} allows for combining validation objects
- * in a way that will allow parsing of nested objects and complex validation rules.
- */
-package io.vavr.control.fluent;
