+++ vavr/src/test/java/io/vavr/control/OptionTest.java
@@ -416,17 +416,17 @@ public class OptionTest extends AbstractValueTest {
 
     @Test
     public void shouldMakeValidOnSomeToValidation() {
-        assertThat(API.Some(5).toValidation("bad")).isEqualTo(API.Valid(5));
+        assertThat(API.Some(5).toValid("bad")).isEqualTo(API.Valid(5));
     }
 
     @Test
     public void shouldMakeLeftOnNoneToValidation() {
-        assertThat(API.None().toValidation("bad")).isEqualTo(API.Invalid("bad"));
+        assertThat(API.None().toValid("bad")).isEqualTo(API.Invalid("bad"));
     }
 
     @Test
     public void shouldMakeLeftOnNoneToValidationSupplier() {
-        assertThat(API.None().toValidation(() -> "bad")).isEqualTo(API.Invalid("bad"));
+        assertThat(API.None().toValid(() -> "bad")).isEqualTo(API.Invalid("bad"));
     }
 
     // -- peek
